+++ src/main/java/nl/dtls/fairdatapoint/api/dto/search/SearchFilterItemDTO.java
@@ -1,0 +1,20 @@
+package nl.dtls.fairdatapoint.api.dto.search;
+
+import lombok.*;
+import nl.dtls.fairdatapoint.api.validator.ValidIri;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Setter
+@Builder(toBuilder = true)
+public class SearchFilterItemDTO {
+
+    @ValidIri
+    private String value;
+
+    private String label = null;
+
+    private boolean preset = true;
+
+}