+++ src/main/java/nl/dtls/fairdatapoint/entity/search/SearchFilterType.java
@@ -1,0 +1,6 @@
+package nl.dtls.fairdatapoint.entity.search;
+
+public enum SearchFilterType {
+    IRI,
+    LITERAL
+}